Move bnsclient to client folder
