+++ cmd/bnsapi/client/bnsclient.go
@@ -1,4 +1,4 @@
-package main
+package client
 
 import (
 	"bytes"
