+++ cmd/bnsapi/client/bnsclient_test.go
@@ -1,4 +1,4 @@
-package main
+package client
 
 import (
 	"bytes"
